[ refact ] : AI 채팅 관련 패키지를 "chatting" 하위 디렉토리로 이동
- AI 채팅 관련 패키지(aiChat, aiSituation)
- userChat 컨트롤러, 서비스 클래스(파일)명 변경
- userChat 엔드포인트 변경: /api/user-chat
