--- src/main/java/com/talkit/app/domain/aiChat/dto/AiChatMessageRequest.java
@@ -1,7 +1,0 @@
-package com.talkit.app.domain.aiChat.dto;
-
-import lombok.Builder;
-
-@Builder
-public record AiChatMessageRequest(String message) {
-}
