--- src/main/java/com/talkit/app/domain/aiChat/repository/AiChatRepository.java
@@ -1,9 +1,0 @@
-package com.talkit.app.domain.aiChat.repository;
-
-import com.talkit.app.domain.aiChat.entity.AiChatRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AiChatRepository extends JpaRepository<AiChatRoom, Long> {
-}
