--- src/main/java/com/talkit/app/domain/aiSituation/repository/AiSituationRepository.java
@@ -1,9 +1,0 @@
-package com.talkit.app.domain.aiSituation.repository;
-
-import com.talkit.app.domain.aiSituation.entity.AiSituation;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AiSituationRepository extends JpaRepository<AiSituation, Long> {
-}
