--- src/main/java/com/talkit/app/domain/aiSituation/service/AiSituationService.java
@@ -1,16 +1,0 @@
-package com.talkit.app.domain.aiSituation.service;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-@Service
-public class AiSituationService {
-
-}
