+++ src/main/java/com/talkit/app/domain/chatting/aiChat/dto/AiChatRoomCreateRequest.java
@@ -1,0 +1,7 @@
+package com.talkit.app.domain.chatting.aiChat.dto;
+
+import lombok.Builder;
+
+@Builder
+public record AiChatRoomCreateRequest(Long situationId) {
+}
