+++ src/main/java/com/talkit/app/domain/chatting/userChat/dto/ChatRoomRequest.java
@@ -1,0 +1,14 @@
+package com.talkit.app.domain.chatting.userChat.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ChatRoomRequest {
+    private String topic; //사용자가 고른 주제
+}
