+++ src/main/java/com/talkit/app/domain/chatting/userChat/entity/ChatMessage.java
@@ -1,0 +1,34 @@
+package com.talkit.app.domain.chatting.userChat.entity;
+
+import com.talkit.app.domain.user.entity.User;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+
+public class ChatMessage {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long cmid; //채팅 식별자
+
+    @ManyToOne
+    @JoinColumn(name = "roomId")
+    private ChatRoom chatRoom;
+
+    @ManyToOne
+    @JoinColumn(name = "senderUid")
+    private User sender; //메세지를 보낸 사용자
+
+    @Column(columnDefinition = "TEXT")
+    private String message; //메세지 내용
+    private LocalDateTime timestamp;
+    private boolean isRead; //읽음여부
+}
